Mention internal flag in api docs
Since this got asked several times it probably makes sense to mention it in the docs for the native api component. To clarify that this is not to be used on the api component but on a per component basis I separated it from configuration variables. Let me know if you have a better idea. I hope the github editor does not break anything.