Added a section for all lambda calls
It was really hard for me to find lambda calls so I made a list. I want to add the light, stepper and servo components to this list but they don't have a heading for lambda instead they have [note: this can also be written in lambda], so I will change those to have a lambda calls heading.