Update homeassistant.rst
Add remark about registration in home assistant.
Even one never plan to export data to home assistant and therefore do not need any entity of the node itself created in home assistant, it is still required to  register the node in home assistant if you want to __import__ data from home assistant into the node. That might not be clear if you come from the "other" side just trying to get data into a node. The paragraph should remind people to do that.