Update st7735.rst
Corrects some mis-statements in the component documentation. It didn't make any sense to say a config item is required and then to say it has a default. It would be better to make those config items optional in the first place.

Also adds the implemented but undocumented "use_bgr" config item.

Thanks for your work in bringing this display controller into esphome.