Added a few extra pointers about device disconnects.
This pull request was earlier on created for describing a new flag that
could be used to disable ACK timeout checking in the AsyncTCP layer for
the API server. However, recent changes in the AsyncTCP code might have
fixed the ACK timeout issues at their root, so I updated the
documentation to suggest an upgrade to 1.18.0 or higher (in which the
AsyncTCP fix is included.)

Also updated the faq.rst file to use "we" instead of "I" wording.