added E-LOG pin on PCB version 2.1 (#903)
Using it with PIR myself, but it is possible to add any type of sensor. The information of the E-LOG pin being GPIO 2, was a guess based on other type of Sonoff devices. I was unable to find information about the pinout of the Sonoff S20 E-LOG pin. This seems like a usefull place to store this information to help others.