Switch the words 'prefix' and 'suffix'
I could be wrong, but I think the words 'prefix' and 'suffix' are used in the wrong place. 
When you append a string, you add text to the end. So the text you add you call a suffix, right?
When you prepend a string, you add text to the beginning. So the text you add you call a prefix, right?