feat: Add manual control config variable doc
This allows to use a time based cover with a manual control switch. The cover state is therefore only a guess so when calling any command (open or close), the action is performed completely nevertheless. At the end of the action, the assumed state is re-synchronized with the cover position.

Fix #2045