Alternative Costco Feit A19 Bulb config (#2591)
I have a Costco Feit A19 that is similar but different. It is actually a RGBCT. I was wondering if we could post it here or otherwise make clear that there are alternate configs for Feit A19 bulbs. Maybe just warning people that there are several models of Feit A19 RGB/white bulbs is enough to not confuse people like me that used the sample code. Also, whoTF makes 2 nearly identical bulbs but with different pinouts WTF is that about, Feit electric?

My change is to the pins of the sm16716 and changing it to an RGBCT bulb.