Clarify what IAQ is
Since standards are only for companies that lack a marketing department, I clarified some of the documentation around Bosch's definition of what IAQ means in their universe.   There is a lot here and it was taken from their datasheet directly that I provided a link to anyway.  I just wanted to include the important stuff to save people from surprises because like me they didn't bother to follow the links.  I can remove the b-VOC stuff if it's too much noise, but I thought it clarified to the users what it might be measuring for.  I also changed `2` to `₂` in `CO2` for the non-code snippets soothe the emotional damage incurred from doing too many school papers in LaTeX.