Added "and/or brightness (e.g. monochromatic)" to the Random Effect description
The Random Effect can also affect the brightness instead of color (e.g. monochromatic light components). I'm not sure if the brightness is also affected for RGB light components, so I put it as an "and/or" instead of just an "or." Of course, adjust based on how it actually works with RGB lights.